requirements.txt
Flask
python-dotenv
google-generativeai
python-pptx
fpdf2
youtube-search-python
google-genai
docx
python-docx
gtts